requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
redis
celery
python-dotenv